Fetch repositories on the server
